Datatables-remove record from movie tbl

+++ Vidly-MVC/Dtos/GenreDto.cs
@@ -7,8 +7,7 @@ namespace Vidly_MVC.Dtos
 {
     public class GenreDto
     {
-        public int Id { get; set; }
-
+        public byte Id { get; set; }
         public string Name { get; set; }
     }
 }
